refactor: migrate MailHistory entity to Kotlin
