+++ src/main/java/com/likelionknu/applyserver/auth/service/UserService.java
@@ -1,8 +1,10 @@
 package com.likelionknu.applyserver.auth.service;
 
 import com.likelionknu.applyserver.application.data.entity.Application;
+import com.likelionknu.applyserver.application.data.entity.MailHistory;
 import com.likelionknu.applyserver.application.data.entity.RecruitAnswer;
 import com.likelionknu.applyserver.application.data.repository.ApplicationRepository;
+import com.likelionknu.applyserver.application.data.repository.MailHistoryRepository;
 import com.likelionknu.applyserver.application.data.repository.RecruitAnswerRepository;
 import com.likelionknu.applyserver.auth.data.dto.request.ModifyProfileRequestDto;
 import com.likelionknu.applyserver.auth.data.dto.response.ProfileResponseDto;
@@ -10,8 +12,6 @@ import com.likelionknu.applyserver.auth.data.entity.Profile;
 import com.likelionknu.applyserver.auth.data.entity.User;
 import com.likelionknu.applyserver.auth.data.repository.UserRepository;
 import com.likelionknu.applyserver.auth.exception.UserNotFoundException;
-import com.likelionknu.applyserver.mail.data.entity.MailHistory;
-import com.likelionknu.applyserver.mail.data.repository.MailHistoryRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
